LeTalk/Backend/requirements.txt
Django==5.1.5
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pymongo==4.11
requests==2.32.3
PyJWT>=2.0
daphne
bcrypt
apscheduler
gunicorn
google-auth
google-auth-oauthlib
boto3
python-dotenv
channels>=4.0.0
channels_redis>=4.1.0

# === TAMBAHKAN di akhir file requirements.txt yang sudah ada ===
httpx
google-generativeai
langchain
langchain-google-genai